Added Class Interface for systemd bootloader spec
Bootloaders implementing the Systemd Bootloader Spec are not
directly compatible with the original Bootloader Class design
in kiwi. Because of that an interface class which translates
the original API into calls following the Systemd Bootloader
Spec was added. This allows us to keep the implementations
in the Builder classes and the public BootLoader interface
untouched. related to Issue #1935