Added overlayroot_verity_setup attribute
Allow to create a dm verity hash placed at the end of the squashfs
compressed read-only root filesystem in an overlayroot setup
The verification credentials information produced by veritysetup
for later verification is stored as a file in /boot/overlayroot.verity.
It is up to the author of the image description how to handle the
credentials file. Without further action the file will be available
in /boot of the image as described above.