some fine tuning
...make sure to squash and update the commit message as follows

Allows to create a dm verity hash, placed at the end
of the squashfs compressed read-only root filesystem.
The verification credentials information produced
by veritysetup for later verification is stored as
a file in /boot/overlayroot.verity and NOT put into
the image by default. It is in the responsibility of the
author of the image description how to handle the
credentials file. One option to handle the credentials
data is as part of a disk.sh script which would allow
to access and handle the information.